requirements.txt
flask==3.1.*
azure-monitor-opentelemetry==1.*
